Removed hash revisions
There was an issue where I was sorting by the revision and hashes were sorted after causing the latest revision to be incorrect. Due to the complexity of grabbing the build date beforehand to sort by date, I have decided to just ignore hash revisions. This may cause some newer revisions to not show, but I think it's better than having to go one by one on all the revisions to grab their information.